antiscam/Backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
scikit-learn>=1.3.2  # Updated to match Colab version (model trained with 1.6.1)
pandas==2.1.4
joblib==1.3.2
numpy==1.26.2
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug==3.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
requests==2.31.0
# tensorflow removed - using sklearn IsolationForest instead

